pawspot/pyproject.toml
[project]
name = "paul-sees-spotify"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "prosodyspeaks", email = "prosodyspeaks@gmail.com" }
]
dependencies = [
    "flask>=3.0.0",
    "gunicorn>=21.2.0",
    "wtforms>=3.1.1",
    "flask_wtf>=1.2.1",
    "spotipy>=2.23.0",
    "flask_session>=0.5.0",
    "flask_bootstrap>=3.3.7.1",
    "python-dotenv>=1.0.0",
    "pawlogger @ git+http://github.com/pawrequest/pawlogger",
    "fastapi>=0.104.1",
    "sqlmodel>=0.0.12",
    "httpx>=0.25.2",
    "pytest-asyncio>=0.21.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ['pytest']

[tool.hatch.metadata]
allow-direct-references = true
